financial-qa/backend/requirements.txt
# Core dependencies — install these first
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
yfinance==0.2.66
curl_cffi>=0.7.0
anthropic>=0.26.1
duckduckgo-search==6.1.7
numpy>=2.0.0
pandas>=2.2.2
pydantic>=2.10.0

# RAG dependencies — install separately when RAG pipeline is ready (see requirements-rag.txt)
# faiss-cpu>=1.9.0
# sentence-transformers>=3.0.0
